fix: add min-height for error message
