+++ src/blocks/form/__button/form__button.css
@@ -1,5 +1,5 @@
 .form__button {
-  margin: 30px 0 0;
+  margin: 15px 0 0;
   padding: 0;
   border-radius: 2px;
   border: 0;
@@ -23,6 +23,6 @@
     font-size: 14px;
     line-height: 17px;
     height: 46px;
-    margin-top: 28px;
+    margin-top: 13px;
   }
 }
